Перевод на typescript node-сервера
