+++ source/server/controllers/devices.ts
@@ -1,8 +1,9 @@
-const _ = require('lodash');
+import _ from 'lodash';
+import * as Express from 'express';
 
-const devicesService = require('../services/devices');
+import devicesService from '../services/devices';
 
-module.exports = function(app) {
+export default (app: Express.Application) => {
   /**
    * @swagger
    * /devices:
@@ -32,7 +33,7 @@ module.exports = function(app) {
    *       200:
    *         description: Массив устройств
    */
-  app.get('/devices', async (req, res) => {
+  app.get('/devices', async (req: Express.Request, res: Express.Response) => {
     const { offset, count, type } = _.get(req, 'query', {});
 
     res.send(await devicesService.getDevices(offset, count, type));
@@ -43,7 +44,7 @@ module.exports = function(app) {
    * /filters:
    *   get:
    *     tags:
-   *        - Filters
+   *        - Devices
    *     description: Возвращает список фильтров
    *     produces:
    *       - application/json
@@ -62,7 +63,7 @@ module.exports = function(app) {
    *       200:
    *         description: Массив фильтров
    */
-  app.get('/filters', async (req, res) => {
+  app.get('/filters', async (req: Express.Request, res: Express.Response) => {
     const { offset, count } = _.get(req, 'query', {});
 
     res.send(await devicesService.getFilters(offset, count));
