+++ source/server/package.json
@@ -1,7 +1,7 @@
 {
   "scripts": {
     "start-streaming-server": "run-p http-server murphy",
-    "start-api-server": "node app.js",
+    "start-api-server": "ts-node app.ts",
     "murphy": "murphy",
     "http-server": "http-server -p 5000 --cors",
     "install-modules": "yarn",
@@ -16,5 +16,15 @@
     "npm-run-all": "^4.1.3",
     "swagger-jsdoc": "^3.2.3",
     "swagger-ui-express": "^4.0.1"
+  },
+  "devDependencies": {
+    "@types/cors": "^2.8.4",
+    "@types/express": "^4.16.0",
+    "@types/lodash": "^4.14.117",
+    "@types/node": "^10.12.0",
+    "@types/swagger-jsdoc": "^0.0.2",
+    "@types/swagger-ui-express": "^3.0.0",
+    "ts-node": "^7.0.1",
+    "typescript": "^3.1.3"
   }
 }
