+++ source/server/tsconfig.json
@@ -1,0 +1,16 @@
+{
+  "compilerOptions": {
+    "target": "esnext",
+    "module": "commonjs",
+    "strict": true,
+//    "moduleResolution": "node",
+    "esModuleInterop": true,
+
+    "outDir": "../../server-dist"
+  },
+
+  "include": [
+    "./**/*.ts"
+  ],
+  "exclude": ["./node_modules/**/*"]
+}
